fix seeding in tests
